# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateTodoRequest:
    properties:
      description:
        maxLength: 500
        minLength: 1
        type: string
      dueDate:
        type: string
      title:
        minLength: 1
        type: string
    required:
    - description
    - dueDate
    - title
    type: object
  dto.UpdateTodoRequest:
    properties:
      description:
        maxLength: 500
        minLength: 1
        type: string
      dueDate:
        type: string
      id:
        type: string
      isCompleted:
        type: boolean
      title:
        minLength: 1
        type: string
    required:
    - description
    - dueDate
    - id
    - isCompleted
    - title
    type: object
  model.Todo:
    properties:
      dateCreted:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: string
      isCompleted:
        type: boolean
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: lluxysa@gmail.com
    name: Murat Can Sahin
  description: This is a simple Todo API server written in Go using MongoDB.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Todo API
  version: 1.0.0
paths:
  /todos:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Todo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: Todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new todo
      tags:
      - todos
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a todo
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: Retrieve a single todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a todo by ID
      tags:
      - todos
    patch:
      consumes:
      - application/json
      description: Update an existing todo item by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a todo
      tags:
      - todos
swagger: "2.0"